Extract/requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
appnope==0.1.3
astroid==2.14.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
blis==0.7.9
boto3==1.26.101
botocore==1.29.76
bs4==0.0.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
confection==0.0.4
confluent-kafka==2.0.2
contourpy==1.0.7
coverage==7.1.0
cycler==0.11.0
cymem==2.0.7
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kaleido==0.2.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.0.0
plotly==5.14.1
plotly-express==0.4.1
pluggy==1.0.0
praw==7.7.0
prawcore==2.3.0
preshed==3.0.8
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.14.0
pylint==2.16.2
pyparsing==3.0.9
pyscopg2==66.0.2
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-mock==1.10.0
rich==13.3.1
s3fs==2023.4.0
s3transfer==0.6.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.9
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
tenacity==8.2.2
textblob==0.17.1
thinc==8.1.10
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
update-checker==0.18.0
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xlrd==2.0.1
yarl==1.8.2
